A Few Final Bugs
1. When running the experiment on other PCs, it fails to retrieve the
images from the file pool. The images appear in the file pool but it
tries to retrieve them from the directory where I used to store the
images on my laptop
2. The batch file doesn't set the subject number correctly
3. At the end of the experiment the highscore and highest_level
variables are presented with the values they had at the start of the
experiment rather than their current values.
4. The high_score variable lags behind by one block